Moved icon and level to base class
